pyproject.toml
[project]
name = "mangalib"
version = "0.1.0"
description = ""
authors = [
    {name = "Darkhan",email = "9heartosu@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
package-mode = false
dependencies = [
    "fastapi[standart] (>=0.116.1,<0.117.0)",
    "uvicorn[standart] (>=0.35.0,<0.36.0)",
    "sqlalchemy[asyncio] (>=2.0.42,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

